Tinder/requirements.txt
wheel==0.26.0
requests
python-dateutil
Flask
flask-ask
